refactor: imports for caching classes
